The output of two testsuite programs has changed.
Intrinsic - A unsupported report is generated for importing a foreign
language entity rather than a warning message.

No_Corresponding_Spec - pragma Global no longer generates an unsuported
report.  The pragma has no effect for ASVAT analysis unless it is an
ASVAT model when it is used to define the global variables of an interface
model.