Remove support for specifying ASVAT models via a pragma Import.
In the CI tests there are many eaxmples of pragma Import with convention
Ada which are not ASVAT models.  It might be possible to resolve such
instances but since Annotate ASVAT is now the preferred way to specify
ASVAT models, it seemed better remove support for specifying via Import.

Also this branch included code to check whether an Intrinsic program had
a body known to gnat2goto.  If this check was to comprehensive it should
also check if an Imported subprogram with convention Ada has body known
to gnat2goto.  After some consideration, it was decided that this
should not be checked by gnat2goto but at the link phase. So the
Intrinsic check has been removed.

Gnat2goto still checks and reports an unsupported node if a called
subprogram is not in the symbol table.